Revert "pint opmaak verbeteringen"
This reverts commit 6399206c34bc83457d0983bf8f30f09a124734cc.

+++ routes/web.php
@@ -1,10 +1,7 @@
 <?php
 
-use App\Http\Controllers\Admin\UserController;
-use App\Http\Controllers\BestellingController;
-use App\Http\Controllers\MateriaalController;
-use App\Http\Controllers\Userzone\ProfileController;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\MateriaalController;
 
 Route::get('/', function () {
     return view('home');
@@ -12,15 +9,15 @@ Route::get('/', function () {
 
 // Admin gebruikersbeheer
 Route::middleware('role:admin')->group(function () {
-    Route::get('/gebruikers', [UserController::class, 'index'])->name('gebruikers');
-    Route::get('/gebruikers/create', [UserController::class, 'create'])->name('gebruikers.create');
-    Route::post('/gebruikers', [UserController::class, 'store'])->name('gebruikers.store');
-    Route::get('/gebruikers/{user}/edit', [UserController::class, 'edit'])->name('gebruikers.edit');
-    Route::put('/gebruikers/{user}', [UserController::class, 'update'])->name('gebruikers.update');
-    Route::delete('/gebruikers/{user}', [UserController::class, 'destroy'])->name('gebruikers.destroy');
+    Route::get('/gebruikers', [\App\Http\Controllers\Admin\UserController::class, 'index'])->name('gebruikers');
+    Route::get('/gebruikers/create', [\App\Http\Controllers\Admin\UserController::class, 'create'])->name('gebruikers.create');
+    Route::post('/gebruikers', [\App\Http\Controllers\Admin\UserController::class, 'store'])->name('gebruikers.store');
+    Route::get('/gebruikers/{user}/edit', [\App\Http\Controllers\Admin\UserController::class, 'edit'])->name('gebruikers.edit');
+    Route::put('/gebruikers/{user}', [\App\Http\Controllers\Admin\UserController::class, 'update'])->name('gebruikers.update');
+    Route::delete('/gebruikers/{user}', [\App\Http\Controllers\Admin\UserController::class, 'destroy'])->name('gebruikers.destroy');
 });
 
-Route::get('/bestellingen', [BestellingController::class, 'index'])
+Route::get('/bestellingen', [\App\Http\Controllers\BestellingController::class, 'index'])
     ->name('bestellingen')
     ->middleware('role:admin,stockbeheerder');
 
@@ -29,16 +26,16 @@ Route::middleware('auth')->group(function () {
     Route::get('/materialen/create', [MateriaalController::class, 'create'])->name('materialen.create');
     Route::post('/materialen', [MateriaalController::class, 'store'])->name('materialen.store');
     Route::view('/winkelmandje', 'pages.winkelmandje')->name('winkelmandje');
-    // Route::view('/bestellingen', 'pages.bestellingen')->name('bestellingen');
+    //Route::view('/bestellingen', 'pages.bestellingen')->name('bestellingen');
     Route::view('/weersvoorspelling', 'pages.weersvoorspelling')->name('weersvoorspelling');
     Route::view('/favorieten', 'pages.favorieten')->name('favorieten');
 
     // Mag ik deze lijn verwijderen??? - Titi
-    // Route::view('/gebruikers', 'pages.gebruikers')->name('gebruikers')->middleware('role:admin');
+   // Route::view('/gebruikers', 'pages.gebruikers')->name('gebruikers')->middleware('role:admin');
 
-    Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
-    Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
-    Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
+    Route::get('/profile', [App\Http\Controllers\Userzone\ProfileController::class, 'edit'])->name('profile.edit');
+    Route::patch('/profile', [App\Http\Controllers\Userzone\ProfileController::class, 'update'])->name('profile.update');
+    Route::delete('/profile', [App\Http\Controllers\Userzone\ProfileController::class, 'destroy'])->name('profile.destroy');
 });
 
 require __DIR__.'/auth.php';
